Reuse Hierarchy's markingClasses() method in SimplifyDataObjectDecorator
SimplifyDataObjectDecorator's markingClasses() had code that was copied from
the Hierarchy extension's method of the same name. This caused trouble when
the original code changed as with SilverStripe 3.1.

Now instead of having to synchronize with Hierarchy with every major version,
we simply call its markingClasses() method first.

Note that simply calling $this->owner->markingClasses() caused segfaults for
me. The use of setOwner() on the returned Hierarchy instance is mandatory,
too, otherwise Hierarchy's other methods won't work correctly.